+++ src/app/[locale]/dashboard/@board/Grid/DroppableCell/DroppableCell.module.css
